feat(crawler): Moving the commands into a more reasonable file structure
